refactored (start/send/receive)TsoApp.ts into a single AddressSpaceApps.ts file
