+++ packages/zostso/src/index.ts
@@ -39,9 +39,7 @@ export * from "./IssueTso";
 export * from "./PingTso";
 export * from "./SendTso";
 export * from "./StartTso";
-export * from "./StartTsoApp";
-export * from "./SendTsoApp";
-export * from "./ReceiveTsoApp";
+export * from "./AddressSpaceApps";
 export * from "./StopTso";
 export * from "./TsoConstants";
 export * from "./TsoResponseService";
